Make SERVER_HOST_URL known to kmp.js.
The file `kpm.js` needs `SERVER_HOST_URL` to be known when processing
it with parcel, so move the parcel processing from a docker build step
to application startup.  Of course, this makes application startup
slower, which may be a problem.

Babel and parcel packages moved from devDependencies to dependencies,
to make them available in the "production" docker image.

The util-linux package in production image is required to get parcel
running.

Temporarily disabled js-cookie in translation.js, as I failed to get
parsel to resolv that.